Changed predicate for checking admin status
The previous predicate checked for admin status far less often. It
only checked when counter was 10,000, state changed, or in the first 5
app cycles. This meant that if there was a misfire in connecting with the
content script, admin mode would not be enabled until much later when it was
checked again. This also removed the error of admin status not being checked if
isAdmin was set to 'null.'