Added 'this course' selection feature
The selectCourses component now fetches the course info for the
course id in the active tab. If a course is not in the active
tab, a message is displayed. This will streamline the process for
selecting a course to scan because the user can go straight to the
course they would like to scan.