Created course scan controllers
Created the CourseScannerController to be used for scanning
individual courses and used by the CoursesScanController, which
is used by the TaskController to start a course scan.

A regex fix for finding course links was also implemented.

The coursesByTermId was added to the CanvasAPIClient and RequestHandler.

The TaskController was implemented in the AppController.