Moved to strategy pattern for scanning courses
Created the Scannable Interface to hold the scannable items of each
course item type. This allowed for a strategy pattern that could handle
the different requirements for the different course item types
without creating overcomplicated code. This improved the code
clarity and maintainability compared to the previous code
used to make the CanvasRequests.