Finished feature for showing results of a single course.
The CourseScanController in the ServiceWorker did not check if there
were any results before adding a course result to the resutls array, so
this was fixed so that the only resultsData would be courses with a match.

The a new CourseItemREsultsSection was added to minimize duplicate code that
was needed for each courseItem type in a scan resutls section.