Added the Security class to increase the reliability and security of the LMS Dash extension.
The Security class introduces a sha256 method to generate hashes
based on input. This is used to create signature for each new message.

The new Message() instantiation must now be followed by the async
setSignature() method to set the message signature. Otherwise, the
message will be rejected when it is passed through the extension.

Each send/receive pattern of CanvasRequests and CanvasResponses must now
be comapred to ensure their message signatures match and each response
has a request with a matching ID.

Each send/receive pattern of messages must now compare the message signatures
to ensure they match.

These changes will help ensure the reliability of the extension.